backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI
openai==1.12.0

# Vector database
qdrant-client==1.7.3

# PostgreSQL
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25

# Document processing
python-markdown==3.5.2
beautifulsoup4==4.12.3
pypdf==4.0.1

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
tiktoken==0.5.2
langchain==0.1.5
langchain-openai==0.0.5

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
